+++ FreecraftCore.Serializer.sln
@@ -46,8 +46,6 @@ Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "FreecraftCore.Serializer.AP
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "FreecraftCore.Serializer.Compression", "src\FreecraftCore.Serializer.Compression\FreecraftCore.Serializer.Compression.xproj", "{DA3E2F6B-4298-4BA0-8D10-D52E786394CA}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "async", "async", "{A0E90E4A-32B5-4135-861D-FCD51741AFA5}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -144,6 +142,5 @@ Global
 		{C4BF70BC-1161-4B69-AB28-2BEDE2765EB0} = {C69D878A-A37C-4DBD-8D90-8AF30074E3C0}
 		{C40130C8-6BFE-4D1F-B27F-EB2DB34BA9B0} = {C69D878A-A37C-4DBD-8D90-8AF30074E3C0}
 		{DA3E2F6B-4298-4BA0-8D10-D52E786394CA} = {C69D878A-A37C-4DBD-8D90-8AF30074E3C0}
-		{A0E90E4A-32B5-4135-861D-FCD51741AFA5} = {C69D878A-A37C-4DBD-8D90-8AF30074E3C0}
 	EndGlobalSection
 EndGlobal
