+++ src/FreecraftCore.Serializer.Stream/EnumerableAsyncBytes.cs
@@ -1,5 +1,4 @@
-﻿#if !NET35
-using System;
+﻿using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
@@ -38,4 +37,3 @@ namespace FreecraftCore.Serializer
 		}
 	}
 }
-#endif
