+++ src/FreecraftCore.Serializer.Stream/Read/Decorators/OneTimeReadReaderStrategy.cs
@@ -11,15 +11,16 @@ namespace FreecraftCore.Serializer
 	/// <summary>
 	/// Decorates a stream the semantics of only being able to read from the stream a single time.
 	/// </summary>
-	public class OneTimeReadReaderStrategy : WireMemberReaderStrategyDecorator
+	public class OneTimeReadReaderStrategy<TReaderType> : WireMemberReaderStrategyDecorator<TReaderType>
+		where TReaderType : IWireStreamReaderStrategy
 	{
 		/// <summary>
 		/// Indicates if the strategy has read yet.
 		/// True if the strategy has yet to be read from.
 		/// </summary>
-		private bool CanRead { get; set; }
+		protected bool CanRead { get; set; }
 
-		public OneTimeReadReaderStrategy([NotNull] IWireStreamReaderStrategy decoratedReader)
+		public OneTimeReadReaderStrategy([NotNull] TReaderType decoratedReader)
 			: base(decoratedReader)
 		{
 			CanRead = true;
@@ -64,58 +65,10 @@ namespace FreecraftCore.Serializer
 			return DecoratedReader.PeakBytes(count);
 		}
 
-		private void ThrowIfCantRead()
+		protected void ThrowIfCantRead()
 		{
 			if(!CanRead)
-				throw new InvalidOperationException($"Cannot read multiple times with {nameof(OneTimeReadReaderStrategy)}.");
+				throw new InvalidOperationException($"Cannot read multiple times with {nameof(OneTimeReadReaderStrategy<TReaderType>)}.");
 		}
-
-		//Async task methods
-#if !NET35
-		/// <inheritdoc />
-		public override Task<byte> ReadByteAsync()
-		{
-			ThrowIfCantRead();
-			CanRead = false;
-
-			return DecoratedReader.ReadByteAsync();
-		}
-
-		/// <inheritdoc />
-		public override Task<byte> PeekByteAsync()
-		{
-			ThrowIfCantRead();
-			CanRead = false;
-
-			return DecoratedReader.PeekByteAsync();
-		}
-
-		/// <inheritdoc />
-		public override Task<byte[]> ReadAllBytesAsync()
-		{
-			ThrowIfCantRead();
-			CanRead = false;
-
-			return DecoratedReader.ReadAllBytesAsync();
-		}
-
-		/// <inheritdoc />
-		public override Task<byte[]> ReadBytesAsync(int count)
-		{
-			ThrowIfCantRead();
-			CanRead = false;
-
-			return DecoratedReader.ReadBytesAsync(count);
-		}
-
-		/// <inheritdoc />
-		public override Task<byte[]> PeakBytesAsync(int count)
-		{
-			ThrowIfCantRead();
-			CanRead = false;
-
-			return DecoratedReader.PeakBytesAsync(count);
-		}
-#endif
 	}
 }
