+++ src/FreecraftCore.Serializer.Stream/Read/Decorators/PrependBytesStreamReaderStrategy.cs
@@ -9,25 +9,26 @@ using JetBrains.Annotations;
 
 namespace FreecraftCore.Serializer
 {
-	public class PrependBytesStreamReaderStrategy : WireMemberReaderStrategyDecorator
+	public class PrependBytesStreamReaderStrategy<TReaderType> : WireMemberReaderStrategyDecorator<TReaderType>
+		where TReaderType : IWireStreamReaderStrategy
 	{
 		/// <summary>
 		/// The prended bytes.
 		/// </summary>
 		[NotNull]
-		private byte[] PrependedBytes { get; }
+		protected byte[] PrependedBytes { get; set; }
 
 		/// <summary>
 		/// The prepended counter.
 		/// </summary>
-		private int ByteCount { get; set; } = 0;
+		protected int ByteCount { get; set; } = 0;
 
 		/// <summary>
 		/// Indicates if the pretended portion is finished reading.
 		/// </summary>
-		private bool isPrendedBytesFinished => (PrependedBytes.Length == ByteCount);
+		protected bool isPrendedBytesFinished => (PrependedBytes.Length == ByteCount);
 
-		public PrependBytesStreamReaderStrategy([NotNull] IWireStreamReaderStrategy decoratedReader, [NotNull] byte[] bytes) 
+		public PrependBytesStreamReaderStrategy([NotNull] TReaderType decoratedReader, [NotNull] byte[] bytes) 
 			: base(decoratedReader)
 		{
 			if (bytes == null) throw new ArgumentNullException(nameof(bytes));
@@ -115,89 +116,5 @@ namespace FreecraftCore.Serializer
 				return bytes;
 			}
 		}
-
-		//Async task methods
-#if !NET35
-		/// <inheritdoc />
-		public override Task<byte> ReadByteAsync()
-		{
-			if (isPrendedBytesFinished)
-				return DecoratedReader.PeekByteAsync();
-			else
-				return Task.FromResult(PrependedBytes[ByteCount++]); //always available because we check length
-		}
-
-		/// <inheritdoc />
-		public override Task<byte> PeekByteAsync()
-		{
-			if (isPrendedBytesFinished)
-				return DecoratedReader.PeekByteAsync();
-			else
-			{
-				return Task.FromResult(PrependedBytes[ByteCount]);
-			}
-		}
-
-		/// <inheritdoc />
-		public override Task<byte[]> ReadAllBytesAsync()
-		{
-			if (isPrendedBytesFinished)
-				return DecoratedReader.ReadAllBytesAsync();
-			else
-			{
-				return new Task<byte[]>(() => PrependedBytes.Skip(ByteCount)
-					.Concat(new EnumerableAsyncBytes(DecoratedReader.ReadAllBytesAsync()))
-					.ToArray());
-			}
-		}
-
-		/// <inheritdoc />
-		public override Task<byte[]> ReadBytesAsync(int count)
-		{
-			if (isPrendedBytesFinished)
-				return DecoratedReader.ReadBytesAsync(count);
-			else
-			{
-				Task<byte[]> bytes = null;
-
-				if (count <= PrependedBytes.Length - ByteCount)
-					bytes = Task.FromResult(PrependedBytes.Skip(ByteCount)
-						.Take(count)
-						.ToArray());
-				else //We need to combine
-					bytes = new Task<byte[]>(() => PrependedBytes.Skip(ByteCount)
-						.Concat(new EnumerableAsyncBytes(DecoratedReader.ReadBytesAsync(count - (PrependedBytes.Length - ByteCount))))
-						.ToArray());
-
-				//Set the byte count as finished or forward as many as possible
-				ByteCount = Math.Min(count + ByteCount, PrependedBytes.Length);
-
-				return bytes;
-			}
-		}
-
-		/// <inheritdoc />
-		public override Task<byte[]> PeakBytesAsync(int count)
-		{
-			if (isPrendedBytesFinished)
-				return Task.FromResult(DecoratedReader.PeakBytes(count));
-			else
-			{
-				Task<byte[]> bytes = null;
-
-				if (count <= PrependedBytes.Length - ByteCount)
-					bytes = Task.FromResult(PrependedBytes.Skip(ByteCount)
-						.Take(count)
-						.ToArray());
-				else //We need to combine
-					bytes = new Task<byte[]>(() => PrependedBytes.Skip(ByteCount)
-						.Concat(new EnumerableAsyncBytes(DecoratedReader.PeakBytesAsync(count - (PrependedBytes.Length - ByteCount))))
-						.ToArray());
-
-				//Peaking so don't move the bytecount forward
-				return bytes;
-			}
-		}
-#endif
 	}
 }
